moved all components out of their own directories and into components directory:
