+++ src/components/BioContainer.tsx
