+++ src/components/index.tsx
@@ -1,6 +1,6 @@
-export * from './bio/BioContainer'
-export * from './contact/ContactContainer'
-export * from './landing/LandingContainer'
-export * from './nav-bar/NavBar'
-export * from './projects/ProjectsContainer'
-export * from './tech-stack/TechStackContainer'
+export * from './BioContainer'
+export * from './ContactContainer'
+export * from './LandingContainer'
+export * from './NavBar'
+export * from './ProjectsContainer'
+export * from './TechStackContainer'
